requirements.txt
nicegui>=2.0
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.0
Pillow>=10.0
imagehash>=4.3
reportlab>=4.0
